fix: do not attempt to share $HOME
If all concurrent Lambda functions share a single $HOME, they end up all
trying to update the same `.npmrc` file. This leads to stale file
descriptor errors (ERRNO -116 / Exit code 140). Instead of sharing the
whole $HOME, configure the NPM cache directory, so that the cache is
shared, but all lambdas keep "their own" home directory.
