+++ src/backend/shared/code-artifact.lambda-shared.ts
@@ -1,5 +1,5 @@
-import { spawn } from 'child_process';
 import { CodeArtifact } from 'aws-sdk';
+import { shellOut } from './shell-out.lambda-shared';
 
 export interface CodeArtifactProps {
   /**
@@ -41,20 +41,4 @@ export async function logInWithCodeArtifact({ endpoint, domain, domainOwner, api
   await shellOut('npm', 'config', 'set', `registry=${endpoint}`);
   await shellOut('npm', 'config', 'set', `${protoRelativeEndpoint}:_authToken=${authorizationToken}`);
   await shellOut('npm', 'config', 'set', `${protoRelativeEndpoint}:always-auth=true`);
-
-  function shellOut(cmd: string, ...args: readonly string[]): Promise<void> {
-    return new Promise<void>((ok, ko) => {
-      const child = spawn(cmd, args, { stdio: ['ignore', 'inherit', 'inherit'] });
-      child.once('error', ko);
-      child.once('close', (code, signal) => {
-        if (code === 0) {
-          return ok();
-        }
-        const reason = code != null
-          ? `exit code ${code}`
-          : `signal ${signal}`;
-        ko(new Error(`Command "${cmd} ${args.join(' ')}" failed with ${reason}`));
-      });
-    });
-  }
 }
